refactor(main.go): change fetchPubMedDetails to use WebEnv and QueryKey
The changes implement the usehistory parameter in the esearch request to
enable the use of WebEnv and QueryKey for fetching PubMed article
details. The fetchPubMedDetails function is refactored to use WebEnv and
QueryKey instead of a list of IDs, which is more efficient and allows
for pagination and more complex queries.