feat(collections): add MapWithError function
Adds a new generic function `MapWithError` to the `collections` package.
This function applies a given function `f` to each element of a slice
`ts`. The function `f` can return a value of type `U` and an error. If
any call to `f` returns an error, `MapWithError` immediately returns
`nil` and the encountered error. Otherwise, it returns a new slice
containing the transformed elements and a `nil` error.